pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "hero-mcp-server"
version = "0.1.0"
description = "MCP Server für HERO Handwerkersoftware"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
hero-mcp-server = "hero_mcp_server.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/hero_mcp_server"]
